Unify chosen geocoding handling
