+++ lib/assets/javascripts/cartodb/common/dialogs/georeference/georeference_view.js
@@ -1,5 +1,6 @@
 var $ = require('jquery');
 var _ = require('underscore');
+var cdb = require('cartodb.js');
 var BaseDialog = require('../../views/base_dialog/view.js');
 var GeoreferenceModel = require('./georeference_model');
 var TabItemView = require('./tab_item_view');
@@ -125,7 +126,7 @@ module.exports = BaseDialog.extend({
   },
 
   _onChangeGeocodeData: function(tabModel, data) {
-    this.trigger('geocodingChosen', data);
+    cdb.god.trigger('geocodingChosen', data);
     this.hide();
   },
