+++ lib/assets/javascripts/cartodb/table/geocoding_dialog/geocoding_dialog.js
@@ -5,7 +5,7 @@
    *  - It only needs a table and a user models.
    *
    */
-  
+
   cdb.admin.GeocodingDialog = cdb.common.CreateDialog.extend({
 
     _TEXTS: {
@@ -121,16 +121,13 @@
         obj.text = true;
       }
 
-      // Check if type is admin and world as location 
+      // Check if type is admin and world as location
       if (obj.kind === "admin1" && isWorld) {
         // Change kind to admin0
         obj.kind = "admin0";
       }
-      
-      // Send trigger to parent
-      this.trigger('geocodingChosen', obj, this);
 
-      // Hide dialog
+      cdb.god.trigger('geocodingChosen', obj);
       this.hide();
     },
