+++ lib/assets/javascripts/cartodb/table/header/options_menu.js
@@ -331,9 +331,6 @@ cdb.admin.HeaderOptionsMenu = cdb.admin.DropdownMenu.extend({
         tabs:   ['lonlat', 'city', 'admin', 'postal', 'ip', 'address'],
         option: 'lonlat'
       });
-
-      dlg.bind('geocodingChosen', this._onGeocodingChosen, this);
-
     } else if (this.options.geocoder.isGeocoding() || ( !canAddGeocoding && !tableIsReadOnly )) {
       if (newModals) {
         dlg = cdb.editor.ViewFactory.createDialogByTemplate('common/background_polling/views/geocodings/geocoding_in_progress');
@@ -348,55 +345,6 @@ cdb.admin.HeaderOptionsMenu = cdb.admin.DropdownMenu.extend({
     dlg
       .appendToBody()
       .open({ center: true });
-  },
-
-  _sendGeocodingMetrics: function(type) {
-    // TODO: remove mixpanel
-    cdb.god.trigger('mixpanel', 'Geocoding', {
-      type: type,
-      table_name: this.table.get('id')
-    });
-
-    // Event tracking "Geocoding"
-    cdb.god.trigger('metrics', 'geocoding', {
-      email: window.user_data.email
-    });
-
-  },
-
-  _onGeocodingChosen: function(data) {
-    this._sendGeocodingMetrics(data.type);
-    var hasNewModals = this.user.featureEnabled('new_modals');
-    var isTypeLonLat = data.type === 'lonlat';
-
-    // Lon-lat type?
-    if (!hasNewModals && isTypeLonLat) {
-      this.table.bind('notice', this.options.globalError.showError, this.options.globalError);
-      this.table.geocodeLatAndLng(data.latitude, data.longitude);
-      return false;
-    }
-
-    if (hasNewModals || !isTypeLonLat) {
-      if (hasNewModals) {
-        var geocodeModel;
-        if (isTypeLonLat) {
-          geocodeModel = new cdb.editor.LonLatGeocodingModel({
-            table: this.table,
-            longitude_column: data.longitude,
-            latitude_column: data.latitude,
-            force_all_rows: !!data.force_all_rows
-          });
-        } else {
-          geocodeModel = new cdb.editor.GeocodingModel(_.omit(data, 'type'))
-        }
-        this.options.backgroundPollingModel.addGeocodingItem(geocodeModel);
-      } else {
-        this.options.geocoder.set(_.omit(data, 'table_name', 'type'));
-      }
-      return false;
-    }
-
-    cdb.log.info('No geocoder option for ' + data.kind, obj);
   },
 
   /**
