+++ lib/assets/javascripts/cartodb/table/noGeoRefData_dialog.js
@@ -52,23 +52,6 @@ cdb.admin.NoGeoRefDataDialog = cdb.admin.BaseDialog.extend({
     return this._TEXTS.content;
   },
 
-  _onGeocodingChosen: function(data) {
-    // Lon-lat type?
-    if (data.type == "lonlat") {
-      this.model.geocodeLatAndLng(data.latitude, data.longitude);
-      return false;
-    }
-
-    // Address, city, region, postal or ip?
-    if (data.type !== "lonlat") {
-      // Remove unnecesary data in any case and start geocoder
-      this.options.geocoder.set(_.omit(data, 'table_name', 'type'));
-      return false;
-    }
-
-    cdb.log.info('No geocoder option for ' + data.type, obj);
-  },
-
   ok: function() {
     if (this.hasContent) {
       var newModals = this.user.featureEnabled('new_modals');
@@ -78,7 +61,6 @@ cdb.admin.NoGeoRefDataDialog = cdb.admin.BaseDialog.extend({
       var canAddGeocoding = bkgPollingModel !== "" ? bkgPollingModel.canAddGeocoding() : true; // With new modals
 
       if (!this.options.geocoder.isGeocoding() && !tableIsReadOnly && canAddGeocoding) {
-
         var GeorefView = this.user.featureEnabled('new_modals') ? cdb.editor.GeoreferenceView : cdb.admin.GeocodingDialog;
         dlg = new GeorefView({
           table:  this.model,
@@ -86,9 +68,6 @@ cdb.admin.NoGeoRefDataDialog = cdb.admin.BaseDialog.extend({
           tabs:   ['lonlat', 'city', 'admin', 'postal', 'ip', 'address'],
           option: 'lonlat'
         });
-
-        dlg.bind('geocodingChosen', this._onGeocodingChosen, this);
-
       } else if (this.options.geocoder.isGeocoding() || ( !canAddGeocoding && !tableIsReadOnly )) {
         if (newModals) {
           dlg = cdb.editor.ViewFactory.createDialogByTemplate('common/background_polling/views/geocodings/geocoding_in_progress');
