+++ lib/assets/javascripts/cartodb/table/table.js
@@ -31,7 +31,7 @@ $(function() {
 
       cdb.admin.hotkeys.enable();
       this.keyBind();
-
+      this._initBinds();
     },
 
     /**
@@ -355,6 +355,57 @@ $(function() {
       this.workView.active(this.workViewActive);
     },
 
+    _initBinds: function() {
+      cdb.god.bind('geocodingChosen', this._onGeocodingChosen, this);
+    },
+
+    _onGeocodingChosen: function(data) {
+      this._sendGeocodingMetrics(data.type);
+      var hasNewModals = this.user.featureEnabled('new_modals');
+      var isTypeLonLat = data.type === 'lonlat';
+
+      // Lon-lat type?
+      if (!hasNewModals && isTypeLonLat) {
+        this.table.bind('notice', this.globalError.showError, this.globalError);
+        this.table.geocodeLatAndLng(data.latitude, data.longitude);
+        return false;
+      }
+
+      if (hasNewModals || !isTypeLonLat) {
+        if (hasNewModals) {
+          var geocodeModel;
+          if (isTypeLonLat) {
+            geocodeModel = new cdb.editor.LonLatGeocodingModel({
+              table: this.table,
+              longitude_column: data.longitude,
+              latitude_column: data.latitude,
+              force_all_rows: !!data.force_all_rows
+            });
+          } else {
+            geocodeModel = new cdb.editor.GeocodingModel(_.omit(data, 'type'))
+          }
+          this.backgroundPollingModel.addGeocodingItem(geocodeModel);
+        } else {
+          this.geocoder.set(_.omit(data, 'table_name', 'type'));
+        }
+        return false;
+      }
+
+      cdb.log.info('No geocoder option for ' + data.kind);
+    },
+
+    _sendGeocodingMetrics: function(type) {
+      // TODO: remove mixpanel
+      cdb.god.trigger('mixpanel', 'Geocoding', {
+        type: type,
+        table_name: this.table.get('id')
+      });
+
+      // Event tracking "Geocoding"
+      cdb.god.trigger('metrics', 'geocoding', {
+        email: window.user_data.email
+      });
+    },
 
     setTable: function(table, sqlView) {
       var self = this;
