+++ lib/assets/javascripts/cartodb/table/tableview.js
@@ -554,8 +554,6 @@
                 data:   { longitude: column }
               });
 
-              dlg.bind('geocodingChosen', this._onGeocodingChosen, this);
-
             } else if (this.options.geocoder.isGeocoding() || ( !canAddGeocoding && !tableIsReadOnly )) {
               if (newModals) {
                 dlg = cdb.editor.ViewFactory.createDialogByTemplate('common/background_polling/views/geocodings/geocoding_in_progress');
@@ -590,30 +588,6 @@
         } else {
           return '<div><div></div></div>';
         }
-      },
-
-      _onGeocodingChosen: function(data) {
-        // Lon-lat type?
-        if (data.type == "lonlat") {
-          this.model.geocodeLatAndLng(data.latitude, data.longitude);
-          return false;
-        }
-
-        // Address, city, region, postal or ip?
-        if (data.type !== "lonlat") {
-          if (this.user.featureEnabled('new_modals')) {
-            // Add new geocoding item
-            this.options.backgroundPollingModel.addGeocodingItem(
-              new cdb.editor.GeocodingModel(_.omit(data, 'type'))
-            )
-          } else {
-            // Remove unnecesary data in any case and start geocoder
-            this.options.geocoder.set(_.omit(data, 'table_name', 'type'));
-          }
-          return false;
-        }
-
-        cdb.log.info('No geocoder option for ' + data.type, obj);
       },
 
 
@@ -833,7 +807,7 @@
           vis: this.options.vis,
           menu: this.options.menu,
           user: this.user,
-          globalError: this.options.globalError,
+          globalError: this.options.globalError
         });
       },
